refactored in new packages for model & parsing. other minor changes.
