+++ src/main/java/ai/humn/telematics/model/JourneySet.scala
@@ -1,14 +1,16 @@
-package ai.humn.telematics
+package ai.humn.telematics.model
 
 /**
  * A JourneySet contains all the journeys for a concrete day and provides some helper methods to compute aggregations
  * on all the journeys (e.g total mileage by driver)
+ *
  * @param journeys Sequence of journeys for a day
  */
 case class JourneySet(journeys: Seq[Journey]) {
 
   /**
    * Computes mileage per driver
+   *
    * @return a Map of driverId -> total mileage
    */
   def mileageByDriver: Map[String, Double] = {
@@ -26,6 +28,7 @@ case class JourneySet(journeys: Seq[Journey]) {
 
   /**
    * Obtains the driver with most kilometers in the JourneySet
+   *
    * @return a tuple containing (driverId, totalMileage) for the most active driver in the JourneySet
    */
   def mostActiveDriver: (String, Double) = {
