+++ src/test/java/ai/humn/telematics/JourneyTest.scala
@@ -1,5 +1,6 @@
 package ai.humn.telematics
 
+import ai.humn.telematics.model.Journey
 import org.junit.runner.RunWith
 import org.scalatest.{FlatSpec, Matchers}
 import org.scalatestplus.junit.JUnitRunner
